requirements.txt
absl-py==2.1.0
asgiref==3.8.1
astunparse==1.6.3
backports.zoneinfo==0.2.1
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
contourpy==1.1.1
cycler==0.12.1
decorator==4.4.2
Django==4.2.16
facenet-pytorch==2.6.0
fer==22.5.1
ffmpeg==1.4
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.54.1
fsspec==2024.10.0
gast==0.4.0
google-auth==2.35.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.67.1
h5py==3.11.0
idna==3.10
imageio==2.35.1
imageio-ffmpeg==0.5.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
Jinja2==3.1.4
keras==2.13.1
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.7.5
moviepy==1.0.3
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
oauthlib==3.2.2
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt_einsum==3.4.0
packaging==24.1
pandas==2.0.3
pillow==10.2.0
proglog==0.1.10
protobuf==4.25.5
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
sqlparse==0.5.1
sympy==1.13.3
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.1
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.4.0
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.6
typing_extensions==4.5.0
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.0.6
wrapt==1.16.0
zipp==3.20.2
